fix: page number error
Former-commit-id: 17542d0c38743b3e0e8c99a33e9e919997b9ea0a

+++ backend/src/main/java/com/jshi/laughtale/chapter/dto/ChapterPaginationDto.java
@@ -8,7 +8,7 @@ public class ChapterPaginationDto {
     @Getter
     @Builder
     public static class Response {
-        private Integer nextPage;
-        private Integer prevPage;
+        private Long nextPage;
+        private Long prevPage;
     }
 }
